pyproject.toml
[tool.poetry]
name = "pythonproject"
version = "0.1.0"
description = ""
authors = ["garickbadalov <garick.badalov@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
youtube-dl = { git = "https://github.com/ytdl-org/youtube-dl.git" }
fastapi = "^0.101.0"
uvicorn = "^0.23.2"
jinja2 = "^3.1.2"
python-multipart = "^0.0.6"
pika = "^1.3.2"
aio-pika = "^9.2.2"
setproctitle = "^1.3.2"
redis = "^5.0.0"
boto3 = "^1.28.36"
yt-dlp = "^2023.7.6"
pytest-playwright = "^0.4.2"
beautifulsoup4 = "^4.12.2"
lxml = "^4.9.3"
minio = "^7.1.16"
aiogram = "3.0"
pydantic = "^2.3.0"
loguru = "0.6.0"
setuptools = "^68.2.2"
pyrogram = {url = "https://github.com/Dineshkarthik/pyrogram/archive/refs/heads/master.zip"}
pyyaml = "5.3.1"
rich = "12.5.1"
tgcrypto = "1.2.5"
werkzeug = "2.2.2"
flask = "2.2.2"
ply = "3.11"
ruamel-yaml = "0.17.21"
flask-login = "0.6.2"
pycryptodome = "3.18.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
